chore: core-questions js to ts refactor (#11617)
* refactor: convert core-questions to TS
